enc/quant.c: silence a warning
score_t -> int: rd_i4.H contains a value from a uint16_t
lookup

+++ src/enc/quant.c
@@ -877,7 +877,7 @@ static int PickBestIntra4(VP8EncIterator* const it, VP8ModeScore* const rd) {
     if (rd_best.score >= rd->score) {
       return 0;
     }
-    total_header_bits += rd_i4.H;   // <- equal to mode_costs[best_mode];
+    total_header_bits += (int)rd_i4.H;   // <- equal to mode_costs[best_mode];
     if (total_header_bits > enc->max_i4_header_bits_) {
       return 0;
     }
